+++ com.unity.render-pipelines.high-definition/Editor/VFXGraph/Outputs/VFXDistortionPlanarPrimitiveOutput.cs
@@ -6,14 +6,26 @@ using UnityEngine;
 namespace UnityEditor.VFX
 {
     [VFXInfo]
-    class VFXDistortionQuadOutput : VFXAbstractDistortionOutput
+    class VFXDistortionPlanarPrimitiveOutput : VFXAbstractDistortionOutput
     {
+        [VFXSetting(VFXSettingAttribute.VisibleFlags.InInspector), SerializeField]
+        protected VFXPrimitiveType primitiveType = VFXPrimitiveType.Quad;
+
         //[VFXSetting] // tmp dont expose as settings atm
         public bool useGeometryShader = false;
 
-        public override string name { get { return "Output Particle Distortion Quad"; } }
-        public override string codeGeneratorTemplate { get { return RenderPipeTemplate("VFXParticleDistortionQuad"); } }
-        public override VFXTaskType taskType { get { return useGeometryShader ? VFXTaskType.ParticlePointOutput : VFXTaskType.ParticleQuadOutput; } }
+        public override string name { get { return "Output Particle Distortion " + primitiveType.ToString(); ; } }
+        public override string codeGeneratorTemplate { get { return RenderPipeTemplate("VFXParticleDistortionPlanarPrimitive"); } }
+        public override VFXTaskType taskType
+        {
+            get
+            {
+                if (useGeometryShader)
+                    return VFXTaskType.ParticlePointOutput;
+
+                return VFXPlanarPrimitiveHelper.GetTaskType(primitiveType);
+            }
+        }
         public override bool supportsUV { get { return true; } }
 
         public override IEnumerable<string> additionalDefines
@@ -26,8 +38,31 @@ namespace UnityEditor.VFX
                 if (useGeometryShader)
                     yield return "USE_GEOMETRY_SHADER";
 
-                yield return "VFX_PRIMITIVE_QUAD";
+                yield return VFXPlanarPrimitiveHelper.GetShaderDefine(primitiveType);
             }
+        }
+
+        protected override IEnumerable<VFXPropertyWithValue> inputProperties
+        {
+            get
+            {
+                IEnumerable<VFXPropertyWithValue> properties = base.inputProperties;
+                foreach (var property in properties)
+                    yield return property;
+
+                if (primitiveType == VFXPrimitiveType.Octagon)
+                    foreach (var property in PropertiesFromType(typeof(VFXPlanarPrimitiveHelper.OctagonInputProperties)))
+                        yield return property;
+            }
+        }
+
+        protected override IEnumerable<VFXNamedExpression> CollectGPUExpressions(IEnumerable<VFXNamedExpression> slotExpressions)
+        {
+            foreach (var exp in base.CollectGPUExpressions(slotExpressions))
+                yield return exp;
+
+            if (primitiveType == VFXPrimitiveType.Octagon)
+                yield return slotExpressions.First(o => o.name == "cropFactor");
         }
 
         public override IEnumerable<VFXAttributeInfo> attributes